Add files via upload
The readFromSensor.ino is what I use to get data from the sensor. HandManipulator.cs is what I use to modify the 3D hand model in Unity. How can I take the data from the ino file and get the readings, convert it into angles, in Unity? I'm basically asking for a solution that connects the two independent codes.